requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
aiosqlite==0.19.0
Pillow==10.1.0